arch: move bootstrap to framework
