+++ packages/cli/package.json
@@ -23,9 +23,6 @@
   },
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
-  "bin": {
-    "ilos-cli": "./dist/cli.js"
-  },
   "mocha": {
     "require": [
       "ts-node/register",
