+++ packages/cli/src/index.ts
@@ -3,11 +3,8 @@ import { Command } from './parents';
 import { CommandRegistry } from './providers';
 import { CommandExtension } from './extensions/CommandExtension';
 import { CliTransport } from './transports/CliTransport';
-import { Bootstrap, bootstrap } from './Bootstrap';
 
 export {
-  Bootstrap,
-  bootstrap,
   CliTransport,
   CommandRegistry,
   Command,
