+++ packages/logger/package.json
@@ -42,9 +42,6 @@
     "extension": [
       ".ts"
     ],
-    "require": [
-      "ts-node/register"
-    ],
     "reporter": [
       "text-summary"
     ],
